* ext/socket: fixed types.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@27529 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/socket/constants.c
@@ -15,7 +15,7 @@ static VALUE rb_mSockConst;
 #include "constdefs.c"
 
 static int
-constant_arg(VALUE arg, int (*str_to_int)(const char*, int, int*), const char *errmsg)
+constant_arg(VALUE arg, int (*str_to_int)(const char*, long, int*), const char *errmsg)
 {
     VALUE tmp;
     char *ptr;
